refactor(platform): drop InProcessDispatch — realign to A3a
A3a is always distributed: edge -> remote DispatchPort -> worker -> run
skill. There is no single-box bot, and the edge never runs skills.
InProcessDispatch invented that topology, mis-named the worker's own
"run a skill" as an edge dispatch, and was built instead of the real
(remote) path. Remove it.

- delete platform/dispatch.py + its test.
- a worker runs skills straight off the registry
  (`registry.get(name).run(...)`); the loop test now proves that path.
- DispatchPort stays in petbot_domain as the edge->worker contract; its
  one concrete impl is remote and ships with the edge (top-down), not as
  an in-process stand-in.

loader/registry/manifest are unchanged — genuine worker (load+run) and
edge (manifest) pieces. All green (ruff, mypy --strict, import-linter,
127 tests).
